requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
demjson==2.2.4
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
